renamed permission page as organization page
